+++ lib/src/extensions/sort.dart
@@ -1,13 +1,31 @@
 import 'package:directed_graph/src/exceptions/error_types.dart';
 import 'package:exception_templates/exception_templates.dart';
 
+/// Extension providing the method [equalTo] for
+/// sorting a set with another set element by element.
+extension CompareSet<T extends Object> on Set<T> {
+  /// Returns `true` if `this` contains the same elements as [other]
+  /// and the elements occur in the same order.
+  bool equalTo(Set<T> other) {
+    final it = iterator;
+    final oit = other.iterator;
+
+    while (it.moveNext() && oit.moveNext()) {
+      if (it.current != oit.current) {
+        return false;
+      }
+    }
+    return true;
+  }
+}
+
 /// Extension providing the method [sort] for
 /// sorting a `Set<T extends Object>` in place.
 extension SortSet<T extends Object> on Set<T> {
   /// Sorts the set in place.
   ///
-  /// * Works with [LinkedHashSet], the default implementation of [Set].
-  /// * Sorting is nonsensical for instances of [HashSet] since
+  /// * Works with `LinkedHashSet`, the default implementation of [Set].
+  /// * Sorting is nonsensical for instances of `HashSet` since
   ///   the iteration order is *not* specified.
   void sort([Comparator<T>? comparator]) {
     if (isEmpty) return;
@@ -25,6 +43,16 @@ extension SortSet<T extends Object> on Set<T> {
     }
     clear();
     addAll(tmp);
+  }
+
+  /// Reverses the element order in place.
+  /// * Works with `LinkedHashSet`, the default implementation of [Set].
+  void reverse() {
+    if (isEmpty) return;
+    if (length == 1) return;
+    final elements = toList().reversed;
+    clear();
+    addAll(elements);
   }
 }